feat: remove event properties from state
