+++ src/Aevatar.Application.Grains/Agents/Creator/CreatorGAgent.cs
@@ -81,24 +81,10 @@ public class CreatorGAgent : GAgentBase<CreatorGAgentState, CreatorAgentGEvent>,
         var eventDescriptionList = new List<EventDescription>();
         foreach (var t in eventTypeList)
         {
-            PropertyInfo[] properties = t.GetProperties(BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static | BindingFlags.DeclaredOnly);
-            var eventPropertyList = new List<EventProperty>();
-            foreach (PropertyInfo property in properties)
-            {
-                var eventProperty = new EventProperty()
-                {
-                    Name = property.Name,
-                    Description = property.GetCustomAttribute<DescriptionAttribute>()?.Description ?? property.Name,
-                    Type = property.PropertyType.ToString()
-                };
-                eventPropertyList.Add(eventProperty);
-            }
-            
             eventDescriptionList.Add(new EventDescription()
             {
                 EventType = t,
-                Description = t.GetCustomAttribute<DescriptionAttribute>()?.Description ?? "No description available",
-                EventProperties = eventPropertyList
+                Description = t.GetCustomAttribute<DescriptionAttribute>()?.Description ?? "No description",
             });
         }
