+++ src/Aevatar.Domain.Shared/Agents/Creator/CreatorGAgentState.cs
@@ -26,13 +26,5 @@ public class EventDescription
 {
     [Id(0)] public Type EventType { get; set; }
     [Id(1)] public string Description { get; set; }
-    [Id(2)] public List<EventProperty> EventProperties { get; set; }
-}
-
-[GenerateSerializer]
-public class EventProperty
-{
-    [Id(0)] public string Name { get; set; }
-    [Id(1)] public string Type { get; set; }
-    [Id(2)] public string Description { get; set; }
+    // [Id(2)] public List<EventProperty> EventProperties { get; set; }
 }
