+++ src/Aevatar.Domain/Subscription/EventTypeDto.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Aevatar.Agents.Creator;
 
 namespace Aevatar.Subscription;
 
@@ -8,4 +7,11 @@ public class EventDescriptionDto
     public string EventType { get; set; }
     public string Description { get; set; }
     public List<EventProperty> EventProperties { get; set; }
+}
+
+public class EventProperty
+{
+    public string Name { get; set; } 
+    public string Type { get; set; }
+    public string Description { get; set; }
 }
